refactor: Edit picker component package
