+++ feature/diary/build.gradle.kts
@@ -10,7 +10,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":feature:mypage"))
     implementation(libs.androidx.appcompat)
     implementation(libs.kotlinx.immutable)
     implementation(libs.retrofit.kotlinx.serialization)
