+++ feature/mypage/build.gradle.kts
@@ -11,6 +11,7 @@ android {
 
 dependencies {
     implementation(project(":feature:auth"))
+    implementation(project(":feature:diary"))
     implementation(libs.retrofit.kotlinx.serialization)
     implementation(libs.lottie.compose)
     implementation(libs.gson)
